ENH: Add classifier and pickled model (#14)
This model uses the same text cleaning method as the original project, but then uses SGDClassifier instead of Perceptron, achieving roughly the same level of accuracy. Unfortunately, dummy classifiers achieve decent performance just because the training dataset is so small and since the 'green' samples are only about 20% of the data. We should get more green samples to add to the training dataset, unless SMEs think that the training datasets class balance reflects reality