CLN: Refactor fbo_nightly_scraper.py and add alternative ftp download method
The new method, read_from_ftp(), relies on ftplib but has the benefit of not needing
to download the file in order to read its contents into memory. This might be beneficial
in a cloud environment where system cleanup (i.e. temp file deletion) occurs while the script is executing.
It might also reduce overhead (but this hasn't been tested)