BUG: vendor textract dependency (#33)
We're vendoring the textract dependency because we needed to manually remove its external dependency on pocketsphinx, which fails to build during its setup.py in cloud.gov. Since we're vendoring a modified version of the textract dependency, we also had to modify textract so that it can run as a script instead of as a module that you import. This meant changing relative imports and rearranging a few files (such as exceptions.py) so that they are directly available for import where needed.