ENH: train new classifiers (#14)
Train three different classifiers and compare their results using 'green' as the positive class, which is also the minority class. Also used the "balanced" version of the class_weight arg for SGDClassifer, which uses the values of y to automatically adjust weights inversely proportional to class frequencies in the input data as n_samples / (n_classes * np.bincount(y)). This resulted in better predictions, where using accuracy as the scorer achieved a roc-auc of .94.