ENH: use SAM notice types for fwd compat (#142)
We'll now be using the SAM notice types of Presoliciation, Solicitation, or Combined Synopsis/Solicitation. We do this to prepare for the SAM migration. We do it by mapping the fbo notice types (e.g. PRESOL) to their SAM counterparts (e.g. Presolicitation). For the FBO MOD type, we have to scrape the proper notice type from the FBO page. If the MOD's base notice type is a sources sought, we now exclude that notice.